Fix id not being set correctly
The problem was that getLastIdCreated() was always returning 0, since nothing has been inserted into the database yet. Even if you call it after inserting, it still returns 0. The solution was to just use the id returned from inserting into the database. Students and Courses can now be safely inserted/deleted.